Simplified open/close menu toggle system
