+++ src/components/crds-shared-header/crds-shared-header.scss
@@ -1,6 +1,6 @@
 // stylelint-disable a11y/no-display-none, a11y/font-size-is-readable
 
-.close {
+.close-nav {
   $padding: 22px;
 
   background: linear-gradient(180deg, rgba(0, 0, 0, 0) 0%, rgba(0, 0, 0, .9) 100%);
