+++ src/components/crds-shared-header/global-nav/global-nav.spec.tsx
@@ -1,6 +1,6 @@
 jest.mock('../../../shared/auth');
 import { GlobalNav } from './global-nav';
-
+//TODO update tests with new toggle system - also include test for fake nav names
 describe('<global-nav>', () => {
   beforeEach(() => {
     this.component = new GlobalNav();
@@ -113,21 +113,25 @@ describe('<global-nav>', () => {
 
   describe('Tests handleProfileClick()', () => {
     it('Checks event stopped if not authenticated', () => {
-      const fakeEvent = {stopPropagation: jest.fn()};
+      const fakeEvent = {
+        stopPropagation: jest.fn()};
 
       this.component.handleProfileClick(fakeEvent);
 
       expect(fakeEvent.stopPropagation).toBeCalled();
     });
 
-    it('Checks navClickHandler to be called if authenticated', () => {
-      const fakeEvent = {};
-      this.component.navClickHandler = jest.fn();
+    it('Checks toggleMenu to be called if authenticated', () => {
+      const fakeEvent = {
+        stopPropagation: jest.fn(),
+        preventDefault: jest.fn()};
+
+      this.component.toggleMenu = jest.fn();
       this.component.auth = {authenticated: true}
 
       this.component.handleProfileClick(fakeEvent);
 
-      expect(this.component.navClickHandler).toBeCalled();
+      expect(this.component.toggleMenu).toBeCalled();
     });
   });
 
@@ -152,7 +156,7 @@ describe('<global-nav>', () => {
   describe('Tests profileClasses()', () => {
     it('Checks nav showing indicator added to class list if nav showing and authenticated', () => {
       this.component.profileNavIsShowing = true;
-      this.component.authenticated = true;
+      this.component.isAuthenticated = true;
 
       const newClass = this.component.profileClasses();
 
@@ -161,7 +165,7 @@ describe('<global-nav>', () => {
 
     it('Checks nav showing indicator not added to class list if nav not showing', () => {
       expect(this.component.mainNavIsShowing).toBeFalsy();
-      this.component.authenticated = true;
+      this.component.isAuthenticated = true;
 
       const newClass = this.component.profileClasses();
 
@@ -170,7 +174,7 @@ describe('<global-nav>', () => {
 
     it('Checks nav showing indicator not added to class list if not authenticated', () => {
       this.component.profileNavIsShowing = true;
-      expect(this.component.authenticated).toBeFalsy();
+      expect(this.component.isAuthenticated).toBeFalsy();
 
       const newClass = this.component.profileClasses();
 
@@ -199,6 +203,229 @@ describe('<global-nav>', () => {
   describe('Tests rootURL()', () => {
     it('should return root URL', () => {
       expect(this.component.rootURL()).toBe('https://int.crossroads.net');
+    });
+  });
+
+  describe('Tests closeNav()', () => {
+    it('Checks navs not showing', () => {
+      const fakeEvent = {
+        preventDefault: jest.fn()
+      };
+
+      this.component.giveNavIsShowing = true;
+
+      this.component.closeNav(fakeEvent);
+
+      expect(this.component.giveNavIsShowing).toBe(false);
+      expect(this.component.mainNavIsShowing).toBe(false);
+      expect(this.component.profileNavIsShowing).toBe(false);
+
+      expect(document.body.style.overflow).toBe('scroll');
+      expect(document.body.style.position).toBe('');
+      expect(document.body.style.width).toBe('');
+    });
+  });
+
+
+  describe('Tests navCloseClasses()', () => {
+    it('Checks expected class name returned when navs are all closed', () => {
+      expect(this.component.mainNavIsShowing).toBe(false);
+      expect(this.component.profileNavIsShowing).toBe(false);
+      expect(this.component.giveNavIsShowing).toBe(false);
+
+      const classes = this.component.navCloseClasses();
+
+      expect(classes).toBe('close-nav');
+    });
+
+    it('Checks expected class name returned when main nav is open', () => {
+      this.component.mainNavIsShowing = true;
+
+      const classes = this.component.navCloseClasses();
+
+      expect(classes).toBe('close-nav is-showing');
+    });
+
+    it('Checks expected class name returned when profile nav is open', () => {
+      this.component.profileNavIsShowing = true;
+
+      const classes = this.component.navCloseClasses();
+
+      expect(classes).toBe('close-nav is-showing');
+    });
+
+    it('Checks expected class name returned when give nav is open', () => {
+      this.component.giveNavIsShowing = true;
+
+      const classes = this.component.navCloseClasses();
+
+      expect(classes).toBe('close-nav is-showing');
+    });
+  });
+
+
+  describe('Tests handleScroll()', () => {
+    beforeEach(() => {
+      this.fakeEvent = {
+        preventDefault: jest.fn()
+      };
+    });
+
+    it('Checks navs closed if main nav showing', () => {
+      this.component.mainNavIsShowing = true;
+
+      this.component.handleScroll(this.fakeEvent);
+
+      expect(this.component.giveNavIsShowing).toBe(false);
+      expect(this.component.mainNavIsShowing).toBe(false);
+      expect(this.component.profileNavIsShowing).toBe(false);
+
+      expect(document.body.style.overflow).toBe('scroll');
+      expect(document.body.style.position).toBe('');
+      expect(document.body.style.width).toBe('');
+    });
+
+    it('Checks navs closed if give nav showing', () => {
+      this.component.giveNavIsShowing = true;
+
+      this.component.handleScroll(this.fakeEvent);
+
+      expect(this.component.giveNavIsShowing).toBe(false);
+      expect(this.component.mainNavIsShowing).toBe(false);
+      expect(this.component.profileNavIsShowing).toBe(false);
+
+      expect(document.body.style.overflow).toBe('scroll');
+      expect(document.body.style.position).toBe('');
+      expect(document.body.style.width).toBe('');
+    });
+
+    it('Checks navs closed if profile nav showing', () => {
+      this.component.profileNavIsShowing = true;
+
+      this.component.handleScroll(this.fakeEvent);
+
+      expect(this.component.giveNavIsShowing).toBe(false);
+      expect(this.component.mainNavIsShowing).toBe(false);
+      expect(this.component.profileNavIsShowing).toBe(false);
+
+      expect(document.body.style.overflow).toBe('scroll');
+      expect(document.body.style.position).toBe('');
+      expect(document.body.style.width).toBe('');
+    });
+
+    it('Checks nothing is changed if no navs are open', () => {
+      this.component.handleScroll(this.fakeEvent);
+
+      expect(this.fakeEvent.preventDefault).not.toBeCalled();
+    });
+  });
+
+  describe('Tests toggleMenu()', () => {
+    beforeEach(() => {
+      this.fakeEvent = {
+        preventDefault: jest.fn(),
+        stopPropagation: jest.fn()
+      }
+    });
+
+    it('Checks toggling main-nav', () => {
+      expect(this.component.giveNavIsShowing).toBe(false);
+      expect(this.component.mainNavIsShowing).toBe(false);
+      expect(this.component.profileNavIsShowing).toBe(false);
+
+      this.component.toggleMenu(this.fakeEvent, 'main-nav');
+
+      expect(this.component.giveNavIsShowing).toBe(false);
+      expect(this.component.mainNavIsShowing).toBe(true);
+      expect(this.component.profileNavIsShowing).toBe(false);
+
+      expect(document.body.style.overflow).toBe('hidden');
+      expect(document.body.style.position).toBe('absolute');
+      expect(document.body.style.width).toBe('100vw');
+    });
+
+    it('Checks toggling main-nav when nav is showing', () => {
+      this.component.mainNavIsShowing = true;
+
+      expect(this.component.giveNavIsShowing).toBe(false);
+      expect(this.component.profileNavIsShowing).toBe(false);
+
+      this.component.toggleMenu(this.fakeEvent, 'main-nav');
+
+      expect(this.component.giveNavIsShowing).toBe(false);
+      expect(this.component.mainNavIsShowing).toBe(false);
+      expect(this.component.profileNavIsShowing).toBe(false);
+
+      expect(document.body.style.overflow).toBe('scroll');
+      expect(document.body.style.position).toBe('');
+      expect(document.body.style.width).toBe('');
+    });
+
+    it('Checks toggling profile-nav', () => {
+      expect(this.component.giveNavIsShowing).toBe(false);
+      expect(this.component.mainNavIsShowing).toBe(false);
+      expect(this.component.profileNavIsShowing).toBe(false);
+
+      this.component.toggleMenu(this.fakeEvent, 'profile-nav');
+
+      expect(this.component.giveNavIsShowing).toBe(false);
+      expect(this.component.mainNavIsShowing).toBe(false);
+      expect(this.component.profileNavIsShowing).toBe(true);
+
+      expect(document.body.style.overflow).toBe('hidden');
+      expect(document.body.style.position).toBe('absolute');
+      expect(document.body.style.width).toBe('100vw');
+    });
+
+    it('Checks toggling profile-nav when nav is showing', () => {
+      this.component.profileNavIsShowing = true;
+
+      expect(this.component.giveNavIsShowing).toBe(false);
+      expect(this.component.mainNavIsShowing).toBe(false);
+
+      this.component.toggleMenu(this.fakeEvent, 'profile-nav');
+
+      expect(this.component.giveNavIsShowing).toBe(false);
+      expect(this.component.mainNavIsShowing).toBe(false);
+      expect(this.component.profileNavIsShowing).toBe(false);
+
+      expect(document.body.style.overflow).toBe('scroll');
+      expect(document.body.style.position).toBe('');
+      expect(document.body.style.width).toBe('');
+    });
+
+
+    it('Checks toggling give-nav', () => {
+      expect(this.component.giveNavIsShowing).toBe(false);
+      expect(this.component.mainNavIsShowing).toBe(false);
+      expect(this.component.profileNavIsShowing).toBe(false);
+
+      this.component.toggleMenu(this.fakeEvent, 'give-nav');
+
+      expect(this.component.giveNavIsShowing).toBe(true);
+      expect(this.component.mainNavIsShowing).toBe(false);
+      expect(this.component.profileNavIsShowing).toBe(false);
+
+      expect(document.body.style.overflow).toBe('hidden');
+      expect(document.body.style.position).toBe('absolute');
+      expect(document.body.style.width).toBe('100vw');
+    });
+
+    it('Checks toggling give-nav when nav is showing', () => {
+      this.component.giveNavIsShowing = true;
+
+      expect(this.component.profileNavIsShowing).toBe(false);
+      expect(this.component.mainNavIsShowing).toBe(false);
+
+      this.component.toggleMenu(this.fakeEvent, 'give-nav');
+
+      expect(this.component.giveNavIsShowing).toBe(false);
+      expect(this.component.mainNavIsShowing).toBe(false);
+      expect(this.component.profileNavIsShowing).toBe(false);
+
+      expect(document.body.style.overflow).toBe('scroll');
+      expect(document.body.style.position).toBe('');
+      expect(document.body.style.width).toBe('');
     });
   });
